layout: save the layout explicitly in Phobos user_md
Instead of saving the binary data in trusted.lov, each element of the
file's layout is read and stored in the user_md. Layout PFL are
supported. Each component is stored in its own entry in the user_md.

Due to a lack of functionnality in Lustre's API, this layout cannot be
restored back. Therefore, the "--restore-lov" option has been removed.

Change-Id: I698646dabb79c73ccd44b26bd213da7c64c35311
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>