bugfix: prevent buffer overflow when processing hints
The size of the user defined hints was not used when
processing hints which resulted in a buffer overflow
because the field hsm_action_item::hai_data is not
null-terminated.

Also, this patch makes the hints API a bit more simple
by intoducing a buffer struct to contain the data and
its size.

Change-Id: I017f8c3788c8d4ed00dc10b13ec83bcab4684baa
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>