build: fix build with gcc 11.4.1
Remove the warning stringop-truncation. This warning triggers errors with
strncpy even when used correctly. Remove the warning instead of
switching to memcpy which would require more careful handling of the
'\0' character which would be more error prone.

Also, use meson dist to build tar file to use the --no-tests option.

Change-Id: I4fd3b330f3d63c6b5b38741ce66f219006e975a0
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>