phobos: update to phobos 2.2
The xfer structure was refactored in a Phobos update, to allow having an
xfer with multiple targets. The code is modified accordingly.

The `alias` option was renamed `profile`. Both are still available as
hints but when using the `alias` a deprecation warning is printed,
asking to switch to `profile`.

The tags structure was mutualized and renamed to string_array. The code
is modified accordingly.

Change-Id: Ia094375773beb768afb611bdb1f0845f3d93efa0
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>